Move __() to helpers.override.php
A separate file is necessary to prevent Laravel from adding __().
